EMPCController/evaluate: much faster execution via Union/feval
Evaluation of single and multiple optimizers is now much faster by resorting
to a new implementation of Union/feval.

Single optimizers are evaluated directly by using the cost as a tiebreak in
Union/feval.

In case of multiple optimizers, we evaluate the cost in each of them and
then pick the one which has the smallest cost.