Polyhedron/contains: complete rewrite
Synopsis
---

Tests whether a point or a polyhedron is a subset of another polyhedron

Syntax
---

status = P.contains(x)
status = P.contains(x, fastbreak)

Inputs:
---

P: polyhedron or an array of polyhedra with "m" elements
x: point or set of "n" points (see note below)
   polyhedron or polyhedron array with "n" elements
fastbreak: boolean, if true, test will be aborted as soon as at least one
           element of "P" contains point "x"
           (false by default)

Note: if "x" is a double, then it must be either a column vector, or a
matrix composed of column vectors. No automatic transposition of "x" is
performed!

Outputs:
---

if "x" is a single point or a single polyhedron:
  "status" is a (m x 1) vector of true/false,
  "status(i)=true" iff "P(i)" contains "x"

if "x" is a matrix of points (i.e., x = [x_1, ..., x_n])
  "status" is a (m x n) matrix of true/false,
  "status(i, j)=true" iff "P(i)" contains point "x(:, j)"

if "x" is a polyhedron, then "P.contains(x)" <=> "x \subseteq P"