Polyhedron/isInside: make non-public, do not check existence of H-rep
Good performance of Polyhedron/isInside() is crucial for fast evaluation of
explicit controllers, therefore the method must be as efficient as possible.

However, just checking whether the input has the H-representation often
consumes majority of the runtime. Since this is ridiculous, we now have the
following:
1) Polyhedron/isInside() is now a non-public method, i.e., does not show up
in "methods Polyhedron"
2) the method implicitly assumes the H-representation is available and does
not explicitly check this

Results in a 2x improved evaluation of explicit controllers.

Use this method only if you know what you are doing! (i.e., you _know_ the
H-representation is available). Prefer Polyhedron/contains whenever
performance is not absolutely critical.

TODO: isInside.c has a slugish performance on complex arrays (e.g. in
test_plcp_21)