PWASystem: store state-update and output functions as PolyUnion
We now store the PWA state-update and output functions in obj.functions as a
PolyUnion object.

This allows to conveniently obtain the state update and the output just by
calling obj.functions.feval([x; u], 'update') and
by obj.functions.feval([x; u], 'output'), respectively. Hence we have a much
simpler implementation of PWASystem/update() and PWASystem/output().

For now we use the first-region tiebreak in feval(). We must ponder what is
the correct way of evaluating discontinuous functions.