PWASystem: store local modes as an array of LTISystem in obj.modes
Individual local modes are now available as an array of LTISystem objects in
the "obj.modes" property. Significantly reduces runtime of PWASystem/toLTI,
which is frequently called during reachability analysis.

Note that a better solution to reduce runtime would be to improve
performance of SystemSignal/copy, but that is not easy.

This change drastically improves runtime of reachability and invariance
computation, see e.g. test_empccontroller_toInvariant_01_pass