Updated documentation for "contains" due modified syntax.
The XML documentation has been updated to match with the modified syntax for "Polyhedron/contains",
"Union/contains", and "PolyUnion/contains".

The method "Polyhedron/contains" has been corrected with an meaningful error message when the
dimension the of point is different.

The test file "test_polyunion_contains_03_pass" has been slightly modified not to display a command output.