AffFunction, QuadFunction: simplify validation in feval()
When evaluating functions via "fun.feval(x)", we now perform just a single
validation: checking that "x" is a column vector of appropriate dimension.

Note that "x" is not even required to be a real vector. This is to allow the
functions to act as penalties, where the penalty needs to be evaluated for
"x" being an sdpvar object during construction of an optimization problem.

Do NOT reject sdpvars in the future or you will break MPC construction!