AffFunction, QuadFunction, NormFunction: new "weight" getter
In preparation of employing Function objects to represent penalties
(see issue #36) we need a consistent way how to obtain the weighting
matrix. This is now done by introducing the "weight" getter:

W = function.weight

* If "function" is QuadFunction, then the getter returns "function.H"
* If "function" is AffFunction, we return "function.F" (the linear term)
* If "function" is NormFunction, we directly store the weight in
function.weight and return it.

In AffFunction, QuadFunction, the "weight" property is Dependent and
Transient, which means it is not stored in the object.