UI: represent penalties by Function objects
This is a big and important change. From now on we are dropping the Penalty
class altogether and use Function objects to model penalties.

Old syntax: obj.penalty = Penalty(Q, type)

New syntax:
* obj.penalty = QuadFunction(Q) if type==2
* obj.penalty = OneNormFunction(Q) if type==1
* obj.penalty = InfNormFunction(Q) if type==Inf
* obj.penalty = AffFunction(Q) if type==0

Basically we can use any Function object as a penalty. In theory, it should
be possible to define nonlinear penalties via Function, as long as YALMIP
can parse it.

The removed Penalty class was replaced by a static function whose purpose is
to notify users about this change. When the old Penalty(Q, type) syntax is
used, the function displays an explanatory warning and automatically returns
a corresponding function representation of the penalty.

mpt_import was modified to return new penalties.

fixes #36