SystemSignal: new "PWApenalty" filter
Adds a PWA penalty on the k-th step of the prediction horizon. The
penalty function must be scalar-valued.

This example adds a PWA penalty on the final predicted state, assumes the
PWA function is given as the function 'obj' of polyunion "PU".

  model.x.with('PWApenalty')
  model.x.PWApenalty.step = N+1; % N+1 is the index of the terminal state
  model.x.PWApenalty.polyunion = PU;
  model.x.PWApenalty.function = 'obj';
  model.x.PWApenalty.isconvex = true;

Note: this filter conflicts with standard "penalty" filters (either
'penalty' or 'terminalPenalty'). It is the user's responsibility to
switch off the standard filters manually (or set the corresponding
standard penalties to zeros).

test_empccontroller_dp_01_pass shows how to solve eMPC problems in a dynamic
programming fashion using this new filter.

fixes issue #51