mpt_plcp: use containers.Map instead of HashTable
The internals of mpt_plcp now use a containers.Map object to represent the
hash table. Gives cleaner implementation and slightly improved overall
runtime.

The outputs of mpt_plcp are the same as before, except of
"sol.xopt.Internal.hash_table_basis", which is now returned as a
containers.Map object (but nobody uses it anyhow).

The HashTable class was removed altogether, since it is no longer required
and used to cause problems when upgrading from MPT2 to MPT3.

Also removing the bogus "hashtbale" (sic) class in the compatibility module.
It was not used anyhow.