AbstractSystem: introduce common update() and output() methods
All classes derived from AbstractSystem now inherit a general implementation
of the update() and output() methods.

Every class which wants to delegate the implementation to AbstractSystem
needs to set the "functions" property to a containers.Map object, which
contains the 'update' and 'output' keys as function handles:

  obj.functions = containers.Map;
  obj.functions('update') = @(x, u) f_up(x, u)
  obj.functions('output') = @(x, u) f_out(x, u)

The particular implementation of "f_up" and "f_out" is up to you, as long as
they take two inputs ("x" and "u") and generate a single output (either the
next state or the system's output).

The output() method reads the "feedthrough" property. You must set it to a
logical, depending on whether the output equation has non-zero terms which
multiply the control input.