SystemSignal/filter_reference: bound the symbolic variable
When symbolic variables are used, we bound them by signal's min/max
values. As a consequence, the optimization problem will be infeasible
if the provided value of the free reference lies outside of signal's
bounds.

The bounding is done primarily for explicit solutions where all
parametric variables need to be bounded.