UI/filters: new format of the "getVariables" callback
Every filters which defines new variables now needs to respond to the
"getVariables feedback by providing a structure (or an array of
structures if there are multiple new variables) of the following
format:

.var: sdpvar representation of the introduced variable
.parametric: logical, set to true if the variable needs to be part of
the vector of initial conditions for the optimization problem; false
otherwise

All variables returned as a response to the "getVariables" filters
will be included to the output of Y=AbstractController.toYALMIP(). In
addition, parametric variables will be part of Y.internal.parameters.

The format of the initial condition can be requested by accessing the
controller.xinitFormat property. The output is a structure:
.names: cell array of names of parametric variables
.components: cell array of names of components which introduced the
respective variable
.dims: dimensions of the respective parametric variables
.n_xinit: required number of initialization variables