FilterBehavior: do not copy/save transient filters
If a filter is configures as transient by setting
FilterSetup.transient=true, then such a filter is neither saved nor
copied.

Declare a filter as transient if it is introduced by other filters
during the addFilter/removeFilter callbacks. Typical example is
the deltaPenalty filter, which requires the initialValue filter to
store the previous value of signal.