SystemSignal: new filter "previous"
The sole purpose of the "previous" filter is to store the value of a
signal obtained during the previous optimization. This is required e.g.
to implement slew-rate penalties and constraints. The filter is
considered internal and is hence hidden from users.

Any filter which requires the previous value of a signal can now access
the previous value in obj.internal_properties.previous.var, provided
it first enables and registers with the "previous" filter (see
SystemSignal/filter_deltaPenalty for an example).

The additional variable required to model the previous value is declared
as a parametric variable which means that it will automatically become
part of the vector of initial conditions.