MPTUIHandle: rename "internal_properties" to "Internal"
For sake of consistency with ConvexSet and Union, we are renaming the
"internal_properties" property of MPTUIHandle and derived classes to
"Internal".

Breaks loading of UI objects saved with previous versions. Providing
backwards compatibility would be possible, but is awkward to implement.
Hence we just throw an error.

For future reference, test_ltisystem_version_1.mat and
test_signal_version_1.mat contain objects saved with a previous version.

fixes issue #62