BinTreePolyUnion: new class to represent binary search trees
Syntax:
  T = BinTreePolyUnion(P)

Inputs:
  P: single PolyUnion object (all sets fully dimensional)
Outputs:
  T: BinTreePolyUnion object

The linear representation of the binary search tree is
available in "T.Tree" as a matrix with the following format:
  T.Tree(i, :) = [hp, index_left, index_right)
    hp: hyperplane to test
    index_left: next row to visit if hp*[x; -1]<=0
    index_right: next row to visit if hp*[x; -1]>0
A negative index represents a leaf node, in which case its
absolute value denotes index of the region which contains the
query point "x".

Additional information about the tree (depth, number of
nodes, structure-based representation of the tree, etc.) is
available in "T.Internal.BinaryTree".

Point location in binary trees is performed by
  [isin, inwhich] = T.contains(x)

Note that the tree discards "outer" boundaries of regions.
Hence T.contains(x) always returns true even if the point
lies outside of the convex hull.

Note that the tree is constructed in a recursive fashion.
Since Matlab imposes an upper limit on the number of
recursive calls, in rare circumstances an error can be
trigger. In that case simply increase the recursion limit and
re-run the tree construction.

fixes issue #34