Fixed bug in Opt class when parsing data from Yalmip.
A small bug has been fixed in Opt class when the data were parsed from Yalmip.  The provided test
"test_opt_23_pass" triggers also another issue that reports the problem to be empty because there is
no degree of freedom for optimization.
The problem was caused in "setMPT26Data" because sometimes the field "qp" was missing in the input
structure and the "Opt" class wrongly associated the problem. This has been fixed in
"Opt/private/setMPT26Data" that checks for presence of "qp" field. If this field is missing, the
type of the problem is determined based on the presence of the quadratic term.