When no MILP/MIQP solver is installed, throw an error message when constructing MILP/MIQP problems.
When no MILP/MIQP solver was installed, the MPTOPTIONS global option handles associated empty
variables "[]" as the names for solvers. This was wrong because the name for the solvers are
accepted only as strings. The problem has been fixed in "mpt_subSolvers" by replacing "[]" with
empty strings ''. The related test file "test_opt_21_pass" has been corrected with the correct type
of the optimization problem.