Fixed wrong computation of the objective value due offset term when solving LP/QP.
When solving parametric problems for a fixed value of the parameters, the offset term "c" in the
objective function  "0.5*u'*H*u + (pF*th+f)'*u + th'*Y*th + C*th + c" was
wrongly calculated in "mpt_solve".  The problem had two sources: 1) the term C*th was not taken into
account and 2) the offset term was not added in the right place (for quick LPs/QPs the problem
was solved earlier before the addition operation was performed). This has been fixed  and tested
with "Opt" tests.