The "solvers" tests have been modified to run without providing arguments.
This change ensures that the top level command "run_all_mpt_tests" does not produce error when
testing the "solvers" directory. Each of the solver tests contains a line that provides the default
arguments and loops thru all installed solvers and sets the tolerance. If something goes wrong, an
error is thrown pointing to which solvers failed.