Corrected Simulink scheme and docs in demos on deployment.
Due to separation of LCP solver to two packages - one for Matlab/Simulink and the other for RTW
extension, the demonstration of online MPC points now to the package LCP solver for RTW. The
simulation can be run on every platform, however, the compilation for XPC target and RTWIN target
can be done only on Windows 32-bit Matlab 2012a. This is due to limitation of Matlab RTW interface
that has keep changing in Matlab for version to version. The new version of Matlab  r2013a does not
support the compilation under RTWIN target because the setup needs to be configured for new CLANG
compiler.

The XML- documentation has been corrected to notify user about the applicability of the demos for
RTW under Windows only.