Fixed problem in Opt class due issue #57.
The parametric problems with the number of decision variables equal to the number of equations (no
degree of freedom) are no longer considered as empty and are solved as in MPT2.6. The error is
thrown only if the number of equalities is greater than the number of decision variables.

This fix spans multiple files:
Opt/validate - the checking for emptyness is done separately  - first for non-parametric problems
and then for parametric problems
Opt/copy - supply more arguments for initial construction of the new object in order not to throw
error due emptyness
Opt/display - the problem with 0 decision variables is considered empty only for non-parametric
problems
Opt/eliminateEquations - fixed bug by renaming the unknown variable "n" to existing S.n"

Tests that check this behavior:
test_opt_23_pass - check construction of problems with fixed degree of freedom
test_opt_31_pass - correction of error message