Updated documentation for "Polyhedron/slice".
Due to modified syntax of "Polyhedron/slice" in changeset 67baea1, the documentation has been
adjusted to output polyhedra that could be in a lower dimension than the original set.
Two examples are provided - one that returns polyhedron in the same dimension and the second that
returns polyhedron in reduced dimension.