Added more tryouts for removing equalities in Opt/eliminateEquations.
The method "Opt/eliminateEquations" has been modified to include more tryouts on finding a
sub-invertible matrix for removing equalities. The change is based on the user feedback [issue162]
"test_opt_eliminateEquations_19_pass" which constructs a problem that is not controllable and hence
the equality constraints build a rank-deficient system.

This particular change loops thru more choices in columns for finding a sub-invertible matrix. If
all choices are wasted, an error is thrown. The provided fail-test actually verifies that required
error message is thrown.