Polyhedron/Polyhedron: Incompatible H- and V-representation is not checked any longer.
If users provide incompatible H- and V-representation, no checking of compatibility is provided and
no error is thrown. It is left completely on the user's responsibility to provide compatible H- and
V-representation. The corresponding piece of code is commented out in "Polyhedron.m" file.

The reason for fixing this problem appeared in Polyhedron/plus method where compatible H- and
V-representation was provided and the result returned by the Polyhedron constructor was incorrect.

Thanks to Davide Raimondo for sending the test file - implemented in "test_polyhedron_40_pass".

The two tests on incompatible H- and V-representation have been modified to be more explanatory, but
they still throw an error indicating that something is wrong.