Polyhedron/project: require all points to be column vectors
If P.project(x) with "x" denoting a single point, then "x" must be
a column vector with "P.Dim" rows.

If the operation is to be performed for multiple points, then "x" must
be a matrix with "P.Dim" rows and "n_points" columns.

Bottom line: don't do automatic transposition. Teach users that points
are column vectors by default.