YSet/project: reject arrays, check correct dimension of "x"
Instead of trying to be smart and doing an automatic transposition, we
now explicitly check that point "x" has the same dimension as
"obj.vars" (which, at least in theory, could be a row vector or a
matrix).

This change also makes YSet/project() to reject arrays. Use forEach()
to invoke the calculation for each element of an array manually.