Union, PolyUnion: introduce the Domain property
All Unions and PolyUnions now have the "Domain" property, which contains
a (simple) representation of the domain. Such a property is useful when
the union is not convex and hence Union/convexHull() would not yield
the desired result.

If the domain is not set, it defaults to Union.Set via the Union/get.Domain
getter. The domain can only be set via the constructor:

  U = PolyUnion('Set', S, 'Domain', D)

If "D" is a polyhedron array, we only retain elements which are not
empty sets.

See issue #82