Polyhedron/buildSetRepresentation: prefer the H-representation
If both the H- as well as the V-representations are available, we now
always prefer the H-rep. The primary purpose of the toolbox is parametric
optimization where polyhedra are generated in H-rep. If both reps are
available, this is usually a consequence of an inadvert H-to-V
conversion, e.g. in plotting. Such a conversion can be unreliable.

As a consequence, the set representation for V-polyhedra is now built
exclusively out of vertices and rays (i.e., Ae*x==be is NOT added).

Supports V-polyhedra without vertices.

Returns formulation of an infeasible problem if the input is an empty
set.