Polyhedron/isPointed: correct wrong reasoning introduced in c25b67ef5f90
It turnes out existence of a vertex is not a sufficient condition to declare
a polyhedron as pointed. Counterexample: a polyhedron with vertices (-1, 0),
(1, 0) and rays (0, 1), (0, -1) is a slab with -1<=x(1)<=1 and x(2) being
arbitrary. It has vertices, yet is not pointed (because it contains a line).

Therefore we revert to the original correct implementation where we perform
the check via H-representation.

We should revisit the implementation if more efficient check exists that
would give the correct answer using only the V-representation.