Revision of "isNeighbor" and "isAdjacent" methods.
The methods have been modified to work with different definitions of adjacency.

The method "isNeighbor" checks if the intersection of two polyhedra is nonempty and of dimension
d-1. This is useful to check whether two polyhedra touch themselves along a facet.

The second method "isAdjacent" is a stricter check which requires that the shared facet belongs to
both of the polyhedra. Basically, the "isAdjacent" method checks if the facet-to-facet property
holds.

The related tests have been modified to cope with the modified functionality of the methods.

Two additional changes have been done in "solvers/mpt_plcp.m" and "PolyUnion/merge". The calls to
these methods has been replaced with the new names of the functions.