mpt_plcp: The graph verification has been extended with more checks.
The subfunction "verify_graph" in the termination part of PLCP solver has been extended with additional checks that
control if the result from the optimization solver is feasible or not. This prevents errors if there
is no point on the facet computed or the point does not lie on the neighboring facet.

The problem was detected when reconstructing explicit MPC design on  a distillation column test case
"http://www.nt.ntnu.no/users/skoge/book/matlab_m/cola/cola.html#matlab" with a reduced model
containing 8 states and horizon 1. More details on the problem formulation are in the ECC13 paper:
Christian Feller, Tor Arne Johansen: "Explicit MPC of higher-order linear processes via
combinatorial multi-parametric quadratic programming".