mpt_plcp: Improved scaling of polyhedra to prevent numerical difficulties.
Based on a report from A. Zanelli, the region-generation part in PLCP solver has been revised to
generate numerically better scaled regions. The implemented change features the scaling algorithm
"mpt_scale_matrix"  that normalizes the matrix A to D1*A*D2 in the system of linear equations as follows:

A*x = b

to

D1*A*D2 y = D1*b

where y = inv(D2)*x

The scaling is implemented for the system of inequalities A*x <= b with the help of diagonal scaling factor
D1>0 to D1*A*x <= D1*b.

The results obtained with the scaled polyhedra seem not to suffer that much from numerical
difficulties as without scaling. However, one can find cases where this approach does not help - for
instance "test_plcp_13_pass".

The related test is stored under "test_plcp_24_pass.m".

In addition to this change, the tolerance factor for determining the neighborhood has been reduced
to 1e-4 from MPTOPTIONS.region_tol. This has been determined empirically based on the "test_plcp_12_pass".