mpt_plcp: Normalization phase shifted at the end of the algorithm.
The scaling that was implemented in the previous commit did not work well with the normalization of
the polyhedra. The problem was detected by Michal with test_plcp_12_pass /gurobi solver.

The normalization phase has been moved out to the end of the PLCP algorithm in order not to conflict
with scaling. This change works fine with GUROBI/LCP/CDD-QUADPROG configuration of solvers.

In addition to this change, the verification of the adjacency list has been modified as well. The
previous detection of neighboring facets (based on set-membership test) has been replaced with
founding a minimum distance to the given facet. This modification is less strict and seems to work
more robustly with the tests.