Extension of code generation for tracking controllers.
The "exportToC" function has been extended with export of explicit controllers supporting tracking
with free reference and delta u formulation. This is a major revision of the function
"exportToC" to allow generation of  a c-code portable in larger projects.

Previously, the generated code was plugged into a single file "mpt_getInput.h" that had to be
included in the evaluation routine. The problem appeared when one wanted to work with multiple
explicit controllers and all generated controllers pointed to a same file. In this change, the
generated code is plugged into a single file and the name of the file is determined by the name of
the variable given by the user. The motivation for this change was a request from S. Mariethoz to be
able to use explicit controllers in larger projects.

The routine "exportToC" generates two files: the routine for evaluation of PWA control law and the
S-function for testing the controller under Simulink. The generated S-function has been tested with
already existing test set. The tests had to be modified to comply with the new syntax.

New demo for demonstration of tracking has been added "mpt_demo_deployment_explicitMPCtracking" and
tested with RTW under win32 platform. The existing demo "mpt_demo_deployment_explicitMPC" has been
corrected.