Unification of code generation methods for PolyUnion/BinTreePolyUnion and EMPCController.
The PolyUnion class contains new method  "toC()" that exports the affine function to C language using
basic sequential search. The method has been derived from existing "BinTreePolyUnion" to support
export of general affine functions.

With this change the method "EMPCController/exportToC()" has been modified to pass the data to
general methods "PolyUnion/toC()" or "BinTreePolyUnion/toC()". The code generating function is
parametrized by the name of the file to be generated such that multiple explicit controllers can be
distinguished. The parametrization applies also for the name of the function that evaluates the PWA
function. The function "exportToC()" generates in addition a C-based S-function that includes the
evaluation PWA function and can be tested within Simulink.

The code generation tests have been modified to support these changes.

The demos on generation of explicit solution did not change, only the Simulink schemes for testing
the controller.

The documentation for this change is not yet provided.