PLCP: Modification of criteria in detection of the closest facet.
The subfunction "verify_graph" in PLCP solver that corrects the adjacency list of the explicit
solution has been modified to improve facet detection of neighboring regions. The test comes into
consideration when the neighborhood check between polytopes P and Q fails but the distance between
then is closer than relative tolerance.

The facet detection is based on generating Chebyshev center xP for a given facet of polytope P and
comparing distance to Chebyshev centre computed for each facet of polytope Q. The points with the
minimum distance are considered as neighboring facets.

This test improves the previous detection that was based on computing distances between xP and each
facet of Q that sometimes resulted in negative values. The new implementation is a slight
modification in  PLCP solver and the performance tested with PLCP test cases seem unchanged.