mpt_plcp: Added fallback scenario for computing the feasible set.
The method for computing the feasible set in the explicit solution was not always able to terminate
with non-empty polytope. The problem originated from numerical difficulties when determining
polyhedra that are very small to tolerances of MPT3. Therefore, a fall-back scenario has been
implemented that checks if the point beyond the given facet gives infeasible PLCP. If it does, then
it is classified as the boundary facet. The fall-back scenario thus involves solving LCP for each
candidate boundary set and therefore is computationally demanding. The fall-back scenario is
activated only if the fast search thru adjacency list fails. In this case the user is informed with
the warning that the adjacency list may be wrong.

Added the test case "test_plcp_25_pass" - as reported from user Stephanie. Note that this test takes
time to terminate because of large number of polyhedra involved.