mpt_solve: Proper parsing of "quicklp" and "quickqp" options.
The function "mpt_solve" did not correctly execute the speed-hack routine. The check was done if the
fields are present in the structure and not again their values. It happened that the speed-hack was
running even when set "quicklp=0" or "quickqp=0". This has  been fixed and the speed-hack is
executed only if "quicklp=1" or "quickqp=1".