PolyUnion: new envelope() method
PolyUnion/envelope() computes the convex envelope of H-polyhedra. The
method is equivalent to MPT2's polytop/envelope() method and implements
the procedure of Section 5.4.3 of the Borrelli, Bemporad, Morari book.

If the union of the polyhedra is convex, then the envelope returns
the convex hull at much lower cost (only LPs need to be solved).

Note that the envelope can be unbounded and, in the worst case, is
equal to R^n. Also note that the returned envelope is in redundant
H-rep.