Polyhedron/computeHRep: use convhulln() by default if possible
convhulln() seems to be more robust than cddmex('hull'). However,
the former is only applicable for bounded, full-dimensional polyhedra
in dimensions >= 2. For all other cases we still need to call cddmex.

Note that convhulln() does not return parameters of the facet. Instead,
it merely calculates which vertices lie on which facet. Once we know
this information, parameters of each facet are determined by solving
a set of linear equations.