mpt_plcp: Graph verification is executed only on request.
The subroutine "verify_graph" does not perform automatic check of the adjacency list in the
postprocessing phase of PLCP solver. The verification can be activated using global option
"MPTOPTIONS.modules.solvers.plcp.adjcheck=true" if required.

The "verify_graph" routine has been disabled because for large partitions in higher dimension the
adjacency check was very time consuming and commercial solvers CPLEX/GUROBI were returning
infeasible problems on distance/chebyCenter problems.

This modification causes that PLCP solver returns flag=-4 in case of wrong adjacency list and the
output PolyUnion object is not marked with "convex" property. The feasible set is wrong and
approximated by a bounding box. User will be informed about this problem by a message that is to be
removed after a fail-back approach for computing a feasible set will be implemented.