Opt/feasibleSet.xml: Added note on computing feasible set.
The note is explained on an example where the feasible set does not cover the union of
regions returned from the parametric solver. It is shown, that in this case the union of regions
represents a subset of the feasible set because the optimality conditions may not hold over entire
feasible set (i.e. cost function can be unbounded).