Fixed exit status on infeasible problems.
This is a minor fix on the parametric solvers MPLP/MPQP because they did not return infeasible
status on infeasible problems. The exit status is now tested against the number of regions stored in
PolyUnion object. In case the PolyUnion does not contain any regions, the exit status is set to
infeasible. One test file has been added "test_opt_solve_03_pass" that checks on this problem for
all parametric solvers.