Opt/qp2lcp: Added conversion from MILP/MIQP to mixed-integer LCP.
Major revision of "Opt/qp2lcp" code to support conversion from MILP/MIQP to mixed-integer LCP. The
transformation algorithm remains the same for continuous variables but for binary and integer
variables are new constraints introduced. For each binary variable there is a complementary variable
that must sum to one. The core of the conversion can be explained on this example:

min 0.5*x'*H*x + (F*th +f)'*x
s.t.  A*x <= b + B*th
           x >= 0
           xB in {0,1}^{nb} - binary variables
           xR in R - real variables

corresponds to LCP

 w - M*z = q + Q*th

[vB]   [             -I              0         0   ][xB ]    [ 1 ]       [ 0  ]
[vR] - [0.5*(Hbr'+Hrb)   Hrr       Ar'  ][xR ] = [ fr ]   + [ Fr ]*th
[ s ]   [             -Ab         -Ar        0  ][lam]    [ b ]       [ B ]

where

vB are the slack variables on every binary xB, vR are the slack variables on real variables xR, s
are the slacks on the inequality constraints and lam are the Lagrange multipliers. The negative
definite part in matrix M steps from the diagonal -I matrix that preserves -1 eigenvalues.

The MILP/MIQP can contain integer variables. In this case, the integer variables are mapped to
binary variables at the time of the construction of the problem. That is, the integer variables xI
are mapped to yB using the affine map xI = T*yB + t that is incorporated into the transformation.

The LCP problem indicates also the types of each variable after conversion which is stored in
"vartype" property.

The backward map between the original primal  x and dual variables is stored in "recover" field
under the map

x = problem.recover.uX*[w; z] + problem.recover.uTh*[th; 1]

To verify that the new mixed-integer conversion works properly, several tests cases have been
provided in "tests/opt" directory. The new tests rely on enumerative PLCP solver (ENUMPLCP) that is
added in this commit. With the introduction of new parametric solver, other files had to be
adjusted. The new solver runs very good under these tests, but more tests on the ENUMPLCP solver
will be provided.