mpt_solvers_options: make defaults compatible with GUROBI 5.6.2
GUROBI 5.6.2 no longer allows values above 2e9 for integer options
(BarIterLimit and SolutionLimit). If the limits should be unbounded, the
options must be set to +Inf.

So far we have used 1e12 for those options, which triggers an error
in 5.6.2. That?s why this patch changes the defaults to Inf.

To be on the safe side, this patch also changes defaults for
certain double options (CutOff, NodeLimit and TimeLimit) from 1e12 to
Inf.