Improved QP2LCP transformation on hybrid MPC problems.
The hybrid MPC  formulations  contain exclusive or condition that is expressed as an equality constraint on
the binary variables (i.e. that only one local model can be valid at a time). This leads to the
formulation that the part of the equality constraint matrix corresponding to the continuous
variables has often rank(A(:,C))=0 and the existing Opt/qp2lcp transformation does not work because
the equality constraints cannot be eliminated.

There are two approaches to resolve this problem:
 1) Model the remaining constraints on binary variables as double-sided inequalities and append
 these to the existing constraints [A; -Ae; Ae]*x <= [b; -be; be] + [pB; -pE; pE]*th
 2) Leave the equality constraints on binary variables and treat them afterwards in the subproblem
 of the parametric solver.

Since the first approach increases the number of inequalities, and consequently the number of
possible combinations in the ENUMPLCP solver , the second approach has been chosen.  For this
reason, the ENUMPLCP solver has been modified to include more equality constraints into the
LP subproblem.

This commit comes with the updated version of ENUMPCP where the stored equality constraints from
hybrid MPC  are directly passed to the LP subproblem. The formulation of the LP subproblem is done
using MPT3 call.

Two more tests "test_enumplcp_03" and "test_enumplcp_04" have been added  to check for this
change. The remaining tests for qp2lcp conversion are in "Opt" directory.