mpt_solve: Added fixing of NaNs.
The legacy implementation of MPQP solver results sometimes in NaNs in the computed control law
because it lacks checks on invertibility of a matrix. The NaNs are further propagated to
polyhedra that are then checked for properties and LP/QP sub-solvers are called.

In this implementation, the NaNs are checked in inequality/equality constraints of "mpt_solve". If the
constraint row contains NaN, it is then replaced with zero row such that the dimension of the
constraints is not changed.

Two tests cases are provided that check NaN in the inequality/equality constraints.