Modification of equality constraints removal.
The equality constraints for a combination of continuous/binary/integer variables cannot be fully
eliminated if the subrank of the equality matrix that corresponds to continuous variables equals
zero. In this case the remaining equality constraints are kept in the output. The error messages in
the previous implementation have been removed and replaced with the message that some of the
equality constraints are still present in the output.

The problem was detected on "test_enumplcp_05" when formulated in YALMIP.  This change required
minor changes in the "qp2lcp" algorithm and "mpt_plcp". In "mpt_plcp" an error is thrown if there
are equality constraints with LCP problem formulation because this problem cannot be solved
(equalities on parameters are not allowed). Instead, the user should specify ENUMPLCP solver to
solve problems with equalities on the parameter.

Two more tests have been added (test_opt_eliminateEquations_46, 47) that transform MIQP to LCP while
keeping some of the equality constraints. Two tests have been modified
(test_opt_eliminateEquations_18, 26).