mpt_solve: check parametric bounds
When Opt.solve(x) is called to solve a parametric problem for a given
value of the parameter, we now check the parametric bounds (Ath*x<=bth).
If the parameter is out of bounds, we report infeasibility.

This patch thus introduces consistency between parametric and numerical
solutions with respect to feasibility for a given parameter.

Fixes issue #108