mpt_plcp: use bounding box instead of projection when feasible set is wrong
If the feasible set is still wrong after fixing the adjacency list, we
now return the outer box approximation of critical regions instead of
resoring to the projection-based Opt/feasibleSet().

The reason is that projections are extremely difficult problems and it
is very likely that the computation will take lots of time. Instead,
we notify the user that the feasible set is wrong and return at least
its approximation.