Polyhedron/projectOnAffineHull: projection on the affine hull
Given a polyhedron P = { x | A*x<=b, Ae*x=be }, calling
Q = P.projectOnAffineHull() produces a new polyhedron
Q = { z | A*F*z <= b - A*x0 } where F is in the null space of
Ae and x0 is any solution to Ae*x0=be.

The equality constraints are identified via
Polyhedron/affineHull().

This method is useful when we need to perform operations on
lower-dimensional sets which are otherwise tricky, such as computing
the volume of a facet or checking the boundedness status, e.g.:

	V = P.projectOnAffineHull().volume()
	s = P.projectOnAffineHull().isBounded()