mpt_plcp: Avoid scaling a matrix with zero columns.
A small fix in the PLCP solver has been implemented to prevent scaling a matrix that contains
zero columns. The scaling routine "mpt_scale_matrix" runs smoothly only if there are no zero rows
ans columns. The PLCP solver contained check for zero rows but the check for zero columns was
missing. This commit adds the check for zero columns and the test is executed  to the data sent by
J. Loefberg "data_test_plcp_27.mat".