Opt/qp2lcp: A minor fix in recovering dual variables.
The issue #114 showed that the dimensions for recovering dual variables did not match in the
transformation procedure PLP/PLCP for a projection problem. This commit fixes the issue #114 by
adjusting the dimension of the matrix for recovering the dual variables for inequality description
"A*x<=b". The data of the projection problem are stored in "data_opt_projtest_qp2lcp46.mat" and the
related test is "test_opt_qp2lcp_46_pass".